+++ Bermuda/Bermuda/DrawableEntity.cpp
@@ -2,7 +2,7 @@
 
 double DrawableEntity::DRAWBUFFER = 64;
 
-DrawableEntity::DrawableEntity(int id, int x, int y, int chunkSize, Image* image)
+DrawableEntity::DrawableEntity(int id, double x, double y, int chunkSize, Image* image)
 	: Entity(id,x,y,chunkSize), drawImage(image)
 {
 	if(image != nullptr)
