+++ Bermuda/Bermuda/Player.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 
 
-Player::Player(int id, double moveSpeed, int x, int y, int chunkSize,Camera* camera, GameStateManager* gsm, MainEntityContainer* mec)
+Player::Player(int id, double moveSpeed, double x, double y, int chunkSize,Camera* camera, GameStateManager* gsm, MainEntityContainer* mec)
 	: Entity(id,x,y,chunkSize), DrawableEntity(id,x,y,chunkSize, nullptr), CollidableEntity(id,x,y,chunkSize), IMovable(moveSpeed)
 {
 	this->mec = mec;
