+++ Bermuda/Bermuda/Rock.cpp
@@ -1,7 +1,7 @@
 #include "Rock.h"
 
 
-Rock::Rock(int id, int x, int y, int chunkSize, MainEntityContainer* mec, Image* rockImage, Image* rockPiecesImage)
+Rock::Rock(int id, double x, double y, int chunkSize, MainEntityContainer* mec, Image* rockImage, Image* rockPiecesImage)
 	: Entity(id,x,y,chunkSize), DrawableEntity(id,x,y,chunkSize, rockImage), CollidableEntity(id,x,y,chunkSize, 0, 14, 32, 18), mec(mec), rockImage(rockImage), rockPiecesImage(rockPiecesImage)
 {
 	mec->getDrawableContainer()->add(this);
